+++ apps/web/package.json
@@ -27,7 +27,6 @@
     "@payloadcms/next-payload": "^0.1.11",
     "@payloadcms/plugin-cloud-storage": "^1.1.1",
     "@payloadcms/richtext-lexical": "^0.7.0",
-    "@payloadcms/richtext-slate": "^1.3.1",
     "@refract-ui/sc": "1.0.1-alpha.38",
     "@storybook/react": "^8.0.4",
     "@svgr/webpack": "^8.1.0",
@@ -36,14 +35,12 @@
     "drizzle-orm": "^0.29.1",
     "escape-html": "^1.0.3",
     "graphql-playground-middleware-express": "^1.7.23",
-    "lodash": "^4.17.21",
     "next": "14.1.4",
     "payload": "^2.4.0",
     "qs": "^6.11.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "sharp": "0.32.6",
-    "slate": "^0.101.5",
     "styled-components": "^6.1.8",
     "vercel-blob-storage": "*"
   },
