+++ apps/web/src/globals/Layout/index.tsx
@@ -4,11 +4,10 @@ import type { PropsWithChildren } from 'react';
 import React, { useMemo } from 'react';
 import Link from 'next/link';
 import type * as themeList from '@mono/theme/src/theme';
+import type { PayloadImageProps } from '@mono/ui/primitives/ResponsivePayloadImage';
 import MaybeThemed from '@mono/web/components/MaybeThemed';
 import ResponsivePayloadImage from '@mono/web/components/ResponsivePayloadImage';
-import expandedDoc from '@mono/web/lib/isExpandedDoc';
 import type { MenuItems, Nav as NavT } from '@mono/web/payload/payload-types';
-import type { PayloadImageProps } from '@mono/web/primitives/primitives';
 import styled, { css } from '@refract-ui/sc';
 import s from 'styled-components';
 
@@ -109,7 +108,7 @@ export interface LayoutType extends PropsWithChildren<NavT> {
 }
 
 function Layout({ children, header, footer, theme }: LayoutType) {
-  const logo = expandedDoc<PayloadImageProps>(header?.logo);
+  const logo = header?.logo as PayloadImageProps;
   const HeaderMenu = useMemo(() => Menu(header?.main), [header?.main]);
   const FooterMenu = useMemo(
     () => Menu(footer?.secondary),
