+++ packages/ui/primitives/PayloadDate.ts
@@ -1,0 +1,3 @@
+export type PayloadDateProps = {
+  pickerAppearance: 'day' | 'month' | 'year';
+};
