+++ packages/ui/primitives/PayloadLink.ts
@@ -1,0 +1,9 @@
+import type { PayloadButtonProps } from './PayloadButton';
+
+export type PayloadLinkProps = {
+  href?: string;
+  text: string;
+  style?: Partial<PayloadButtonProps>;
+  newTab?: boolean;
+  linkType?: string;
+};
